# perf: replace varargs in hot paths to make the code jittable