--- a/gateway/src/apicast/executor.lua
+++ b/gateway/src/apicast/executor.lua
@@ -21,9 +21,9 @@ local mt = { __index = _M }
 
 -- forward all policy methods to the policy chain
 for _,phase in Policy.phases() do
-    _M[phase] = function(self, ...)
+    _M[phase] = function(self)
         ngx.log(ngx.DEBUG, 'executor phase: ', phase)
-        return self.policy_chain[phase](self.policy_chain, self:context(phase), ...)
+        return self.policy_chain[phase](self.policy_chain, self:context(phase))
     end
 end