@davidor 2018-10-18
I updated the commit and the PR comments with a more reliable test that isolates the impact that the echo endpoint might have on the results.

The code is the same.